% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIcon.R
\name{CIcon}
\alias{CIcon}
\title{Returns final CI}
\usage{
CIcon(deltahat, deltaSigma, Al, Au, c_LF, alphac, tol, tol_r, g)
}
\arguments{
\item{deltahat}{(B x k) matrix}

\item{deltaSigma}{(B x k) matrix}

\item{Al}{constraint matrices}

\item{Au}{constraint matrices}

\item{c_LF}{critical value}

\item{alphac}{adjusted alpha level}

\item{tol}{tolerance for correlation threshold}

\item{tol_r}{tolerance for correlation threshold}

\item{g}{mapping function returning lambda estimates}
}
\value{
Return final CI
}
\description{
Returns final CI
}
