% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIcon_TNbounds.R
\name{CIcon_TNbounds}
\alias{CIcon_TNbounds}
\title{Computes list(th_1, th_2): lower / upper bounds vectors}
\usage{
CIcon_TNbounds(
  theta,
  deltahat,
  Al,
  Au,
  sigma_l,
  sigma_u,
  corr_m,
  corr_l,
  corr_u,
  cLFl,
  cLFu,
  tol_r,
  g
)
}
\arguments{
\item{theta}{scalar}

\item{deltahat}{(B x k) matrix}

\item{Al}{constraint matrice}

\item{Au}{constraint matrices}

\item{sigma_l}{standard deviations of lambda_hat_l / u}

\item{sigma_u}{standard deviations of lambda_hat_l / u}

\item{corr_m}{correlation blocks}

\item{corr_l}{correlation blocks}

\item{corr_u}{correlation blocks}

\item{cLFl}{critical values}

\item{cLFu}{critical values}

\item{tol_r}{tolerance for correlation threshold}

\item{g}{mapping function returning lambda estimates}
}
\value{
list(th_1, th_2): lower / upper bounds vectors
}
\description{
Computes list(th_1, th_2): lower / upper bounds vectors
}
