% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIproj_p.R
\name{CIproj_p}
\alias{CIproj_p}
\title{# Return final projection indicator}
\usage{
CIproj_p(
  c,
  delta,
  alphac,
  c_bd,
  deltastar_demean,
  Al,
  Au,
  sigma_l,
  sigma_u,
  deltaSigma,
  corr_m,
  corr_l,
  corr_u,
  eta,
  tol_r,
  g
)
}
\arguments{
\item{c}{modifier}

\item{delta, }{d row or column vectors/matrices.}

\item{alphac}{adjusted critical level}

\item{c_bd}{critical value}

\item{deltastar_demean}{row or column vectors/matrices.}

\item{Al}{constraint matrices}

\item{Au}{constraint matrices}

\item{sigma_l}{standard deviations of lambda_hat_l / u}

\item{sigma_u}{standard deviations of lambda_hat_l / u}

\item{deltaSigma}{(B x k) matrix}

\item{corr_m}{correlation blocks}

\item{corr_l}{correlation blocks}

\item{corr_u}{correlation blocks}

\item{eta}{critical area level}

\item{tol_r}{tolerance for correlation threshold}

\item{g}{mapping function returning lambda estimates}
}
\value{
Return final projection indicator
}
\description{
# Return final projection indicator
}
