% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cov_corr_boot.R
\name{estimate_cov_corr_boot}
\alias{estimate_cov_corr_boot}
\title{Estimates bootstrap covariance of correlation vector}
\usage{
estimate_cov_corr_boot(x, y, z, B = 800, seed = 123)
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector.}

\item{z}{Numeric vector.}

\item{B}{Integer number of bootstrap replications.}

\item{seed}{Integer seed for reproducibility.}
}
\value{
A 3x3 covariance matrix.
}
\description{
Estimates bootstrap covariance of correlation vector
}
