% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_rho.R
\name{estimate_rho}
\alias{estimate_rho}
\title{Estimates rho vector from data}
\usage{
estimate_rho(data, x_col = "x", y_col = "y", z_col = "z")
}
\arguments{
\item{data}{Data frame.}

\item{x_col}{Name of x column.}

\item{y_col}{Name of y column.}

\item{z_col}{Name of z column.}
}
\value{
Named numeric vector of correlations.
}
\description{
Estimates rho vector from data
}
