% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_g_on_grid.R
\name{eval_g_on_grid}
\alias{eval_g_on_grid}
\title{Evaluates g function on a grid}
\usage{
eval_g_on_grid(g_fun, rho_vec, grid_r_xu)
}
\arguments{
\item{g_fun}{Function mapping rho and r_xu to lambda.}

\item{rho_vec}{Named numeric vector of correlations.}

\item{grid_r_xu}{Numeric vector of grid points.}
}
\value{
Numeric vector of g values.
}
\description{
Evaluates g function on a grid
}
