% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_fun.R
\name{g_fun}
\alias{g_fun}
\title{Return value of r_zu}
\usage{
g_fun(r, rho_xy, rho_xz, rho_yz)
}
\arguments{
\item{r}{r_xu values}

\item{rho_xy}{observed correlation values}

\item{rho_xz}{observed correlation values}

\item{rho_yz}{observed correlation values}
}
\value{
Return value of r_zu
}
\description{
Return value of r_zu
}
