% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_grad.R
\name{g_grad}
\alias{g_grad}
\title{Return value of gradient of g function}
\usage{
g_grad(r, rho_xy, rho_xz, rho_yz)
}
\arguments{
\item{r}{r_xu values}

\item{rho_xy}{observed correlation values}

\item{rho_xz}{observed correlation values}

\item{rho_yz}{observed correlation values}
}
\value{
Return value of gradient of g function
}
\description{
Return value of gradient of g function
}
