% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_xu_safe.R
\name{g_xu_safe}
\alias{g_xu_safe}
\title{Computes the CIs and probabilities}
\usage{
g_xu_safe(r_xu, delta, tol = 1e-10)
}
\arguments{
\item{r_xu}{Numeric scalar, value of correlation coefficient rho_xu.}

\item{delta}{Numeric vector for observed correlations coefficient: r_xy, r_xz, r_yz.}

\item{tol}{tolerance level}
}
\value{
values of function g(rho, r_xu).
}
\description{
Computes the CIs and probabilities
}
