% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivcrtest-package.R
\docType{package}
\name{ivcrtest-package}
\alias{ivcrtest}
\alias{ivcrtest-package}
\title{ivcrtest: Correlation-Restriction Test for Instrument Validity}
\description{
Tests whether the orthogonality restriction on an instrument remains
compatible with the data once the sign and magnitude of endogeneity are
restricted.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Ainura-tur/ivcrtest}
  \item Report bugs at \url{https://github.com/Ainura-tur/ivcrtest/issues}
}

}
\author{
\strong{Maintainer}: Ratbek Dzhumashev \email{ratbek.dzhumashev@monash.edu}

Authors:
\itemize{
  \item Ratbek Dzhumashev \email{ratbek.dzhumashev@monash.edu}
  \item Ainura Tursunalieva \email{ainura.tursunalieva@csiro.au}
}

}
\keyword{internal}
