% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_compute_gradient_safe.R
\name{local_compute_gradient_safe}
\alias{local_compute_gradient_safe}
\title{Computes gradient of g_xu(r_xu, rho_xy, rho_xz, rho_yz)}
\usage{
local_compute_gradient_safe(r_xu, rho_xy, rho_xz, rho_yz, tol = 1e-10)
}
\arguments{
\item{r_xu}{assumed correlation  values.}

\item{rho_xy}{observed correlation values.}

\item{rho_xz}{observed correlation values.}

\item{rho_yz}{observed correlation values.}

\item{tol}{Numeric scalar for error tolerance.}
}
\value{
3 x length(r_xu) matrix (each column is a gradient vector).
}
\description{
Computes gradient of g_xu(r_xu, rho_xy, rho_xz, rho_yz)
}
