% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_cr_test.R
\name{print.iv_cr_test}
\alias{print.iv_cr_test}
\title{Print method for CR test results}
\usage{
\method{print}{iv_cr_test}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"iv_cr_test"}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for CR test results
}
