% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_c_im.R
\name{solve_c_im}
\alias{solve_c_im}
\title{Solves for Imbens-Manski critical value}
\usage{
solve_c_im(alpha, t)
}
\arguments{
\item{alpha}{Numeric significance level.}

\item{t}{Numeric ratio of identified set width to max SE.}
}
\value{
Numeric critical value.
}
\description{
Solves for Imbens-Manski critical value
}
