% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_compatibility.R
\name{check_compatibility}
\alias{check_compatibility}
\title{Computes the CIs and probabilities}
\usage{
check_compatibility(df, i = 1, n, k = 1, alpha = 0.05, rxu_range = rxu_range)
}
\arguments{
\item{n}{Numeric scalar for size of the variables.}

\item{k}{Numeric scalar for direction of endogenity.}

\item{alpha}{Numeric scalar for the level of H0 test.}

\item{rxu_range}{Numeric interval for r_xu values.}

\item{data}{Data frame}

\item{seed}{Numeric scalar  for reproducibility.}

\item{bias_mc}{Logial value.}

\item{mc_B}{Numeric scalar for MC correction simulations.}
}
\value{
CIs and probabilities.
}
\description{
Computes the CIs and probabilities
}
