% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_simple_union.R
\name{ci_simple_union}
\alias{ci_simple_union}
\title{Computes the CIs and probabilities}
\usage{
ci_simple_union(x, y, z, rxu_range = c(0, 0.8), alpha = 0.05)
}
\arguments{
\item{x}{Numeric vector (endogenous variable).}

\item{y}{Numeric vector (outcome variable).}

\item{z}{Numeric vector (Instrumental variable).}

\item{rxu_range}{Numeric interval for r_xu values.}

\item{alpha}{Numeric scalar for the level of H0 test.}
}
\value{
Simple CIs.
}
\description{
Computes the CIs and probabilities
}
