% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIcon_TNbounds.R
\name{CIcon_TNbounds}
\alias{CIcon_TNbounds}
\title{Computes list(th_1, th_2): lower / upper bounds vectors}
\usage{
CIcon_TNbounds(
  theta,
  deltahat,
  Al,
  Au,
  sigma_l,
  sigma_u,
  corr_m,
  corr_l,
  corr_u,
  cLFl,
  cLFu,
  tol_r,
  g
)
}
\arguments{
\item{theta}{Scalar}

\item{deltahat}{(B x k) matrix}

\item{Al}{Lower constraint matrix}

\item{Au}{Upper constraint matrix}

\item{sigma_l}{Standard deviations of lambda_hat_l}

\item{sigma_u}{Standard deviations of lambda_hat_u}

\item{corr_m}{Correlation matrix for TS statistic}

\item{corr_l}{Correlation matrix for lower bound}

\item{corr_u}{Correlation matrix for upper bound}

\item{cLFl}{Critical value for lower}

\item{cLFu}{Critical value for upper}

\item{tol_r}{Tolerance for correlation}

\item{g}{Mapping function g(delta)}
}
\value{
list(th_1, th_2)
}
\description{
Computes list(th_1, th_2): lower / upper bounds vectors
}
