% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIhybrid.R
\name{CIhybrid}
\alias{CIhybrid}
\title{Returns final confidence intervals}
\usage{
CIhybrid(
  deltahat,
  deltaSigma,
  Al,
  Au,
  alpha,
  alphac,
  eta,
  B,
  Blarge,
  tol,
  tol_r,
  index,
  g
)
}
\arguments{
\item{deltahat}{row or column vectors/matrices.}

\item{deltaSigma}{(B x k) matrix}

\item{Al}{constraint matrices}

\item{Au}{constraint matrices}

\item{alpha, }{critical levels}

\item{alphac}{critical levels}

\item{eta}{critical area level}

\item{B, }{number of repetitions for bootstrap}

\item{Blarge}{number of repetitions for bootstrap}

\item{tol}{tolerance for correlation threshold}

\item{tol_r}{tolerance for correlation threshold}

\item{index}{index number}

\item{g}{mapping function returning lambda estimates}

\item{sigma_l}{standard deviations of lambda_hat_l / u}

\item{sigma_u}{standard deviations of lambda_hat_l / u}
}
\value{
Return final confidence intervals
}
\description{
Returns final confidence intervals
}
