% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_cr_test.R
\name{iv_cr_test}
\alias{iv_cr_test}
\title{Perform Instrumental Variable Correlation Restirction test}
\usage{
iv_cr_test(
  data,
  X,
  Y,
  H,
  Z,
  n = NULL,
  k = -1,
  alpha = 0.05,
  seed = 123,
  rxu_range = c(0, 0.8),
  bias_mc = TRUE,
  mc_B = 500
)
}
\arguments{
\item{data}{A data frame.}

\item{X}{Name of the endogenous variable.}

\item{Y}{Name of the dependent variable.}

\item{H}{Vector of exogenous variable names.}

\item{Z}{Name of the instrumental variable.}

\item{n}{Numeric scalar for size of the variables.}

\item{k}{Numeric scalar for direction of endogenity.}

\item{alpha}{Numeric scalar for the level of H0 test.}

\item{seed}{Numeric scalar  for reproducibility.}

\item{rxu_range}{Numeric interval for r_xu values.}

\item{bias_mc}{Logical value.}

\item{mc_B}{Numeric scalar for MC correction simulations.}

\item{reps}{Number of bootstrap loops.}
}
\value{
A list containing CIs and coverage or containment probabilities.
}
\description{
Perform Instrumental Variable Correlation Restirction test
}
\examples{
df <- wooldridge::card  # Use sample data set
data <- df[complete.cases(df), ]  # Remove missing values
attach(data)
H0 <- data.frame(lwage, educ,fatheduc, exper,expersq ,black ,south ,smsa ,smsa66,reg661, reg662 ,reg663, reg664,
reg665, reg666 ,reg667,reg668)
Y <- as.character(colnames(H0))[1] ###OUTCOME variable
X <-as.character(colnames(H0))[2] ###ENDOEGENOUS variable
Z <-as.character(colnames(H0))[3] ###IV variable
H <-as.character(colnames(H0))[-(1:3)] ##### EXOGENOUS variables
result <- iv_cr_test(data,X,Y,H,Z,  n, k=-1,alpha = 0.05,seed = 123,rxu_range = c(0, 0.8),bias_mc = TRUE,mc_B = 500)
print(result)
}
